Add CLI command to ignore transpilation of certain files/folders (#3)
# New Functionality

## Transpilation
Transpiliation can be performed while bundling by adding the `-t` or `--transpile` flag to the cli command:
```
npx njs-bundle --transpile
```

The transpiler is configured using the `babel-preset-njs` library which is tuned to only transpile code that is not compatible with njs.  Even so *you should aim to transpile as little as possible, or ideally not at all* since it adds complexity to both build, maintenance, and debugging.

### Modes
Transpilation is performed in three different modes:
* `NONE` - Nothing is transpiled
* `SCOPED` (default) - The tool attempts to only transpile code that you did not write.  By default, it assumes that your njs source code is at `**/src/**/*.?(m)js`.  However, you can also pass a glob to specify the pattern to skip transpilation by passing the `-i` or `--transpile-ignore` flag. See the next subsection for more details.
* `ALL` - All code will be run through the transpiler

### Scoping Transpilation
When operating in `SCOPED` mode (which is the default if the `--transpile` flag is passed) you can additionally pass the `-i` or `--transpile-ignore` flag.

This flag takes a single "glob" type argument that specifies the pattern that should be used when *skipping* transpilation.  For example, if I know that I will be writing my njs scripts using only njs-compatible syntax and I keep them in `/code/scripts` then I could specify:

```
npx njs-bundle -t SCOPED -i '/code/scripts/*.?(m)js'`
```

By default, the tool assumes anything in `**/src/**/*.?(m)js` should not be transpiled - so if you don't want to have to configure this every time just put your njs scripts under a `src` directory.

# Structural/technical changes
## Removed virtual filesystem for tests
Since the tests run a process that invokes module resolution, there were a lot of complexities in setting up the virtual filesystem to play nicely with ESM module resolution.  The support for ESM is still experimental in node technically so a lot of the libraries are either not supporting it or early in their support.  Thus, cross library incompatibility was introducing too much complexity.

So we use the real filesystem and clean up afterwards.  This is done with the acknowledgement that this is:
* Slow
* Prone to error (files could be left on the system)
* Prevents parallel test runs (as currently written)
However, for this small project with a limited number of tests, the speed tradeoff seems ok, and the complexity of managing files in a `tmp` folder is nothing compared to the headache of the issues with the virtual filesystems and ESM module resolution. 

Ideally in the future we revisit this and do it properly with mocks and the like - but for now this accomplishes the goal

## Rename of `DEPENDENCIES` transpile mode to `SCOPED`
Since Dependencies wasn't technically accurate after adding the globbing mechanic with `-i`, I changed the name to not be as misleading, since you could opt to transpile things that are not dependencies if so desired.